doc-upload-function/requirements.txt
functions-framework
google-cloud-storage
Flask
Werkzeug